# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VGuardrail Audit Cloud API
  version: "0.1.0"
  description: >-
    Agent-facing API for the VGuardrail AI Governance Platform. Supports the
    end-to-end agent flow: device registration, signed-policy download, and
    tamper-evident audit-event ingestion.

    Audit records form a per-device hash chain (event_hash / previous_event_hash)
    and ingestion is idempotent on event_id and upload_id.
  license:
    name: Apache-2.0
servers:
  - url: https://{api_id}.execute-api.{region}.amazonaws.com
    variables:
      api_id: { default: "REPLACE" }
      region: { default: "us-east-1" }
...
paths:
...
  /policies/latest:
    get:
      tags: [agent]
      summary: Download the latest signed policy bundle for the caller's org.
      description: >-
        Returns the Ed25519-signed bundle the engine verifies and loads.
        Supports conditional GET: send If-None-Match with the current version to
        receive 304 when unchanged.
      operationId: getLatestPolicy
      security:
        - cognitoJWT: []
      parameters:
        - in: header
          name: If-None-Match
          required: false
          schema: { type: string }
          description: ETag (policy version) the client currently holds.
      responses:
        "200":
          description: The latest signed policy bundle.
          headers:
            ETag:
              schema: { type: string }
              description: Current policy version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyBundle" }
        "304":
          description: Not modified (client already holds the latest version).
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No policy published for the org yet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    cognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito access token (device user). Validated by the API Gateway JWT authorizer.
    enrollmentSecret:
      type: apiKey
      in: header
      name: X-Enrollment-Token
      description: Per-org enrollment secret (Secrets Manager). Register only.
  responses:
...
    Unauthorized:
      description: Missing/invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    PolicyBundle:
      type: object
      description: >-
        Ed25519-signed policy bundle (the engine's unit of distribution). The
        cloud stores and serves it; it never signs (offline/CI signing).
      required: [schema, version, org_id, created_at, default_action, rules, signature]
      properties:
        schema: { type: string, example: "vguardrail.policy/v1" }
        version: { type: integer }
        previous_version: { type: integer, nullable: true }
        org_id: { type: string }
        created_at: { type: string, format: date-time }
        default_action: { type: string, enum: [allow, warn, block] }
        exceptions: { type: array, items: { type: object, additionalProperties: true } }
        rules: { type: array, items: { type: object, additionalProperties: true } }
        signature:
          type: object
          required: [alg, key_id, value]
          properties:
            alg: { type: string, const: ed25519 }
            key_id: { type: string }
            value: { type: string, description: base64 Ed25519 signature. }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }